Switch to PSR-4 namespace and file arrangement Rename class namespaces more appropriately Remove all `require` statements Make phpunit use autoloader
