+++ lib/Consumer/Consumer.php
