+++ lib/Consumer/File.php
