+++ lib/Consumer/QueueConsumer.php
