+++ lib/Consumer/Socket.php
